feat: refact chart component
