+++ apps/website/src/components/chart/components/index.tsx
@@ -1,0 +1,5 @@
+export { Areas } from './areas'
+export { ChartComponent } from './chart-component'
+export { ChartTooltip } from './chart-tooltip'
+export { Lines } from './lines'
+export { Markers } from './markers'
