+++ apps/website/src/components/chart/components/lines.tsx
@@ -1,0 +1,77 @@
+import { animated } from '@react-spring/web'
+import { curveMonotoneX } from '@visx/curve'
+import { LinePath } from '@visx/shape'
+
+import type { SpringValue } from '@react-spring/web'
+import type { ScaleLinear, ScaleTime } from 'd3-scale'
+
+type Datum = {
+  date: Date
+  value: number
+}
+
+type Props = {
+  colors: {
+    total: string
+    closed: string
+    background: string
+  }
+  yScale: ScaleLinear<number, number, never>
+  xScale: ScaleTime<number, number, never>
+  totalIssuesData: Datum[]
+  closedIssuesData: Datum[]
+  drawingLineStyle: {
+    strokeDasharray: SpringValue<string>
+  }
+}
+
+const AnimatedLinePath = animated(LinePath)
+
+const Lines = (props: Props) => {
+  const {
+    closedIssuesData,
+    colors,
+    drawingLineStyle,
+    xScale,
+    yScale,
+    totalIssuesData,
+  } = props
+  return (
+    <>
+      {/* Total issues line */}
+      <AnimatedLinePath
+        data={totalIssuesData}
+        x={d => {
+          const datum = d as Datum
+          return xScale(datum.date)
+        }}
+        y={d => {
+          const datum = d as Datum
+          return yScale(datum.value)
+        }}
+        stroke={colors.total}
+        strokeWidth={2}
+        curve={curveMonotoneX}
+        style={drawingLineStyle}
+      />
+
+      {/* Closed issues line */}
+      <AnimatedLinePath
+        data={closedIssuesData}
+        x={d => {
+          const datum = d as Datum
+          return xScale(datum.date)
+        }}
+        y={d => {
+          const datum = d as Datum
+          return yScale(datum.value)
+        }}
+        stroke={colors.closed}
+        strokeWidth={2}
+        curve={curveMonotoneX}
+        style={drawingLineStyle}
+      />
+    </>
+  )
+}
+export { Lines }
