+++ apps/website/src/components/chart/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useAnimations } from './use-animations'
+export { useChartTooltip } from './use-chart-tooltip'
